forum_proj/requirements.txt
asgiref==3.10.0
Django==5.2.7
dotenv==0.9.9
gunicorn==23.0.0
mariadb==1.1.14
mysqlclient==2.2.7
packaging==25.0
python-dotenv==1.2.1
sqlparse==0.5.3
whitenoise==6.11.0
django-axes==8.0.0
python-json-logger==2.0.7